requirements.txt
django
google-api-python-client
google-auth
google-auth-oauthlib
pandas
matplotlib